Develop

See merge request onti-2019-fintech/fintech2019003/pay-by-face!18

+++ tools.py
@@ -1,6 +1,16 @@
+try:
+    import dlib
+except:
+    pass
+
 from eth_account import Account
 from math import ceil
 from json import load
+
+try:
+    from imutils import face_utils
+except:
+    pass
 from web3 import Web3, HTTPProvider
 import uuid
 import sha3
@@ -8,6 +18,10 @@ import cognitive_face as cf
 import os
 import time
 import cv2
+
+
+from random import choice, randint
+
 
 
 def parceJson(file_path):
@@ -58,7 +72,14 @@ def get_balance_by_priv(web3, priv):
 
 def clear(to):
     for i in range(1, to + 1):
-        os.remove(str(i) + ".jpg")
+        if os.path.exists(str(i) + ".jpg"):
+            os.remove(str(i) + ".jpg")
+
+
+def clear_files(names):
+    for name in names:
+        if os.path.exists(name):
+            os.remove(name)
 
 
 def exist_group(create=False):
@@ -89,6 +110,7 @@ def check_all_right(func=cf.person_group.lists, *args, **kwargs):
             elif e.status_code == 429:
                 if hasattr(e, 'message'):
                     time.sleep(int(e.message.split("Try again in ")[1].split()[0]))
+                    return check_all_right(func, *args, **kwargs)
         elif hasattr(e, 'code'):
             if e.code == 5:
                 exit()
@@ -135,6 +157,100 @@ def create_frames_simple(video):
         return face_ids
     else:
         return False
+
+
+def get_open_eyes(frame):
+    eyes = [False, False]
+    gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
+    detector = dlib.get_frontal_face_detector()
+    predictor = dlib.shape_predictor("./opt/shape_predictor_68_face_landmarks.dat")
+    rects = detector(gray, 0)
+    eye_ar_thresh = 0.2
+    for ind, rect in enumerate(rects):
+        shape = predictor(gray, rect)
+        shape = face_utils.shape_to_np(shape)
+        (lStart, lEnd) = face_utils.FACIAL_LANDMARKS_IDXS["left_eye"]
+        (rStart, rEnd) = face_utils.FACIAL_LANDMARKS_IDXS["right_eye"]
+        left_eye = shape[lStart:lEnd]
+        right_eye = shape[rStart:rEnd]
+        left_ear = eye_aspect_ratio(left_eye)
+        right_ear = eye_aspect_ratio(right_eye)
+        if right_ear < eye_ar_thresh:
+            eyes[1] = True
+        if left_ear < eye_ar_thresh:
+            eyes[0] = True
+    return eyes
+
+
+def get_open_mouth(frame):
+    gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
+    detector = dlib.get_frontal_face_detector()
+    predictor = dlib.shape_predictor("./opt/shape_predictor_68_face_landmarks.dat")
+    rects = detector(gray, 0)
+    eye_ar_thresh = 0.4
+    for ind, rect in enumerate(rects):
+        shape = predictor(gray, rect)
+        shape = face_utils.shape_to_np(shape)
+        (lStart, lEnd) = face_utils.FACIAL_LANDMARKS_IDXS["mouth"]
+        lStart += 12
+        mouth_dots = shape[lStart:lEnd]
+        mar = mouth_aspect_ratio(mouth_dots)
+        if mar > eye_ar_thresh:
+            return True
+    return False
+
+
+def mouth_aspect_ratio(mouth):
+    a = euclidean(mouth[1], mouth[7])
+    b = euclidean(mouth[2], mouth[6])
+    c = euclidean(mouth[3], mouth[5])
+    d = euclidean(mouth[0], mouth[4])
+
+    mar = (a + b + c) / (2.0 * d)
+
+    return mar
+
+
+def eye_aspect_ratio(eye):
+    a = euclidean(eye[1], eye[5])
+    b = euclidean(eye[2], eye[4])
+    c = euclidean(eye[0], eye[3])
+
+    ear = (a + b) / (2.0 * c)
+
+    return ear
+
+
+def check_right_rotation(image, right_rotation, max_error, type=0):
+    check = {0: ["roll", "yaw"], 1: ["roll"], 2: ["yaw"]}
+    res = check_all_right(cf.face.detect, image, attributes="headPose")
+    rig = 0
+    if res:
+        rot = res[0]["faceAttributes"]["headPose"]
+        for ch in check[type]:
+            for right in right_rotation:
+                if type == 0:
+                    if (right - max_error) <= rot[ch] <= (right + max_error):
+                        pass
+                    else:
+                        return False
+                else:
+                    if (right - max_error) <= rot[ch] <= (right + max_error):
+                        rig = right
+                        return str(rig)
+        if type == 0:
+            return str(rig)
+        else:
+            return False
+    else:
+        return False
+
+
+def euclidean(p, q):
+    sum_sq = 0.0
+    for i in range(len(p)):
+        sum_sq += (p[i] - q[i])**2
+    return sum_sq ** 0.5
 
 
 def get_predict(faces):
